chore(infra): test Node.js v16 and expand "engines" range to include it (#713)
* infra(ci): Switch to using `cimg/` Docker images over `circleci/` images

The newer images are smaller and suggested to be widely cached.  In
practice, the older images are widely cached because they're still more
prevalent but happy to use the newer images, particularly since the older
images aren't getting as much in the way of updates anymore (e.g., there are
not any Node.js 16 images for it.)

* docs: Switch to using `NODE_VERSION` and `NPM_VERSION` for Netlify

Using this instead of `.nvmrc` since this allows more granular control over
the `npm` version that gets installed.

Ref: https://docs.netlify.com/configure-builds/manage-dependencies/#npm

* infra: Start running Node.js 16 tests

* chore: bump upper-bounds of `node` within `engines` to allow Node.js v16

* Remove unnecessary `engines` constraint on unpublished package

The constraint on this can be enforced exclusively at the monorepo level
since this package is not published and that constraint need not be conveyed
by any user at runtime.

* renovate: Intentionally disable the `circleci` manager

The "circleci" manager is intentionally disabled as of this commit.

While we do benefit from its updating of "Orb" versions, the CircleCI
manager also attempts to update Docker images using its "Docker" datasource.

This is really handy, in theory, but we have intentionally varying major
Node.js docker image identifiers in our CircleCI configuration to test on
each major Node.js platform!  I suspect that enabling the "circleci" manager
would cause all of these to update to the latest version (e.g., 16) when we
want them to be intentionally different!

I'm going to try to craft an fix for this, but I might do it upstream on
Renovate itself using its Regex manager.  It's also worth noting that there
is other configuration that could be used to _only_ renovate "orbs" on this
file, but I'm taking the short-cut route and not juggling with that right
now.  I can test the repository locally with my own copy of Renovate and
come up with another solution given some free time.