chore: adjust pattern to align with upstream example
This introduces the `op_sync` function, which automatically de-serializes
the first ZeroCopyBuf and serializes the return value to reduce boilerplate.

Honestly, this isn't the best implementation for this, but I think we can
change the use of our invocation of the `op_print` to instead use something
that is more like `op_compose` and actually doesn't do the buffer draining
itself?

I think this can be worked on, but this does at least align with the example.

Ref: https://github.com/denoland/deno/commit/fec1b2a5a4324a7e