Add locations to composition errors (#686)
This commit adds location information to composition errors that are specific to federation. This includes pre, post and composition errors. The locations provided on composition errors point to the most relevant location in SDL where the error occurs.

For now, there is no way to map these locations back to the correct subgraph. This problem will be addressed in later unit of work (see #705).

Co-authored-by: Jesse Rosenberger <git@jro.cc>
Co-authored-by: Trevor Scheer <trevor.scheer@gmail.com>