Fix printing of `@join_field`
The changes we made to include applied directives on the field metadata broke a hacky workaround we used to include `@join_field(graph:)` on fields of the owning service, and our fix of that broke `@provides` on fields of value types. This commit cleans up the code a bit so we replicate the previous behavior in a more straightforward way.