Implemention of query-graphs and composition
"Query graphs" (probably not the best name, need to find something
better) are graphs build from graphQL schemas. At a high level, it's
vertex are the named types of the schema and the edge are fields or
downcast (edges between an interface an one of its implementation for
instance). For subgraphs, we construct a specific version of such
"query graph": it's a single graph, that includes each subgraph "query
graph" as a subgraph, but contains additional edges and vertices to
represent federation information like @key, @provides and @requires.

This representation is the base on which both composition validation
and query planning is then built:
- composition validation ensure we can match any traversal on the
  supergraph "query graph" by an equivalent traversal on the
  federated subgraph "query graph".
- query planning takes a query that is valid on the supergraph,
  and find all traversals of the federated subgraph "query graph"
  that match it (composition validation essentially guarantees
  there is at least one such traversal). It then pick whichever
  traversal is most efficient and build the plan from it.

This commit include the code for building/navigating the "query graph"
abstraction, and the first of it's application, composition validation.
It also include the "merge" part of composition that takes the subgrpah
schema and creates the supergraph schema (compostion validation goals
being to tell use if such merging is "usable" in practice).