gateway: Update gateway.load() to support AS2 and AS3
Starting with AS 2.25.0, gateway.load receives `graphRef` in addition to
`graphId` and `graphVariant`; `graphVariant` is declared as always
provided.

In AS3, only `graphRef` will be provided; `graphVariant` will not be
provided.

In order for this code to work with both versions, we declare a version
of ApolloConfig that could come from either ourselves, and parse it into
an internal representation based on the AS3 version.