Add OpenTelemetry support
This PR adds instrumentation to emit the following spans:

federation.request - total time to server a request.
federation.validate - time to validate the federated query.
federation.plan - time to plan the federated query.
federation.execute - the total time executing query.
federation.fetch - time fetching data from a subgraph.
federation.postprocessing - time to render the response from all the subgraph data.
The instrumentation looks a little messy due to the requirement to set error status on spans on exception, but also if errors are added to the response.

It is a reissue of the following PRs
https://github.com/apollographql/federation/pull/803
https://github.com/apollographql/federation/pull/828