Update approach - disallow @tag and @external together
This reverts most of the previous implementation in favor of a new
composition error which prevents users from confusingly using both
@tag and @external together. For now, @tag is expected to have
implications with schema contracts and in the context of an external
field, it's not well-defined what that would actually mean. We may
revisit and relax this in the future, but for now have decided to
just prevent the possible confusion.