Remove `@inaccessible` elements when converting to API schema (#807)
This is an initial implementation of support for @inaccessible when converting
a composed schema to an API schema. For now, it removes any fields, object
types, or interface types with an @inaccessible directive.

As part of this change, buildComposedSchema no longer returns an API
schema. Instead, a separate toAPISchema function is used to convert the
composed schema. The reason for this is that the gateway needs the complete
schema for query planning, including elements that will be removed by
@inaccessible.

Co-authored-by: Trevor Scheer <trevor.scheer@gmail.com>
Co-authored-by: Jesse Rosenberger <git@jro.cc>