fix(deps): Narrow `graphql` peerDependencies to versions that actually work (#913)
Narrow `graphql` peer dependency to a more fitting range `^15.4.0` based on our
current usage of the package. This requirement was introduced by, but not
captured in, the most recently released pre-1.0 minor release of the federation
packages. As such, this change will be released as a `patch` since the breaking
change already accidentally happened and this is a correction to that oversight.