Consider only the most direct paths from a subgraph to another
If we had 3 subgraphs, A, B and C, with key edges for a type T between
all of them, when computing which states we could go from T in A, we
were considering the path A -> B and A -> C, but also A -> B -> C and
A -> C -> B. The latter ones are equivalent to the former ones and never
preferable, so this commit ignore them.

Note that considering those "useless" states was, in practice, adding a
lot of processing (when a type is spread on more than 3 subgraph, the
state reduction is even more dramatic) so this path lower the time to
validate composition on the integration schema from ~500 milliseconds
("on my machine") to ~50 milliseconds.