Caches 'non-collecting' paths from a given vertex
When an entity type (in a subgraph) can be reached from more than one
path (pretty common), validation and query-planning were recomputing
multiple times the set of non-collecting paths that can be obtained from
that type, even though the options were always the same (that is, the
path leading to the type was different, but the set of non-collecting
edges we could take from that point is always the same).

This commit "fixes" this by caching the information for each vertex
once it's been computed the first, saving recomputations.

"On my machine", this improves the validation time for the integration
tests schema from ~50 milliseconds to ~40 milliseconds.