Fix test that is now more optimized
The 'should preserve type conditions for value types' test had been
added in the fix for #396: that fix was primarily ensure that we
did not elimiate type condition, even when they not strictly useful
to the rest of the selection, _if_ they have directives we need
to preserve. But that particular test was actually testing the
case where the type condition has no directives and is thus,
strictly speaking, useless, and the new query planner now optimize
that case "properly".

Note that the next test, 'should preserve directives on inline fragments
even if the fragment is otherwise useless', test the more interesting
where we do want to preserve a type condition due to a directive,
and this test is (still) properly handled.