Fix tests where the order of fields has changed in the query
The new query planning code sometimes does not query fields in exactly the same
order than the the previous one (at least for fields that are not part of the
user query but added by the planner, like `__typename` or key fields).

It's unfortunate that the test require a particular order of fields
since it's irrelevant to correct behavior, but as that is the case, this
commit changes the expected order in the few cases where it changed.