Revert "Gateway: pass through errors encountered when running final execute (#159)"
This reverts commit 9045f553075829393c8dfd9a69898701178d7b72, which was from
https://github.com/apollographql/federation/pull/159.

See the comment in https://github.com/apollographql/federation/issues/981
which explains the justification as well, but briefly:

While that PR was an improvement in principle, it:

- has a bug that was first captured in
  https://github.com/apollographql/federation/pull/159#issuecomment-901246314
To quote that here:

  > I think there's a bug here. Let's say we have a query `{ x }` where `x`
is a non-nullable field, and the federated execution has `x` throw an error.
This turns all of `data` into `null` (not `{ x: null }`). But then this
re-execution adds another error saying that the non-null field is null.
  >
  > Take a look at
https://codesandbox.io/s/angry-raman-uuzuf?file=/src/index.js for an example
of what's going on here.

- Sometimes — unexpectedly to current users, at least — breaks client
  expectations.

As of now, the pain points seem to be outweighing the gains.  We'll need to
revert this and revisit this when time allows with a slightly different
approach and https://github.com/apollographql/federation/issues/981 tracks
the need to do so.

Ref: https://github.com/apollographql/federation/pull/159
Ref: https://github.com/apollographql/apollo-server/issues/5550
Ref: https://github.com/apollographql/federation/issues/974
Ref: https://github.com/apollographql/apollo-server/pull/4523