Fix composition validation with "unsatisfiable" branches
Composition validation can get into a situation where one can
express a type condition `... on T` on the supergraph after an
operation, but that the supergraph handling that operation cannot
ever return a `T` (because it doesn't even know that type).
In that case, composition was failing, but it shouldn't: instead
if should say that this is fine, it is simply a "branch" that will
always yield no results if requested (and we can always trivally plan
a query that return no results after all).