Improves composition performance
The composition validation algorithm was mistakenly re-validating the same
vertices/edges of the supergraph _a lot_ (once for every possible path from a root going
through that edge, which for large schema can explode for some edges). But this
is unecessary and was not the intent: if we can validate that we can reach
everything from a particular vertex from one incoming path, then we can do the
same from any incoming path and so the idea is that we (mostly) don't have to validate
vertices/edges multiple times.

In practice, it is a little bit more subtle as when we reach a vertex (a type)
in the supergraph, then the actual incoming path might have got us in only a
subset of the subgraphs having that type. Another incoming path that reaches that
same vertex may get us there with a different subset of subgraphs. So in practice,
we remember for each vertex the same of subgraphs we were in when reaching it
and only avoid re-validating that vertex if the "current" set of subgraphs we're in
is the same as in a previous visit of that vertex. If it isn't, we do
re-validate (and remember that new set of subgraphs). Meaning that a supergraph
vertex may be re-validated multiple times, but only so much time.

This commit mostly implements this optimization. Without it, composition validation
essentially explode exponentially as the schema grows. With it, each vertex/edge
of the supergraph is only handled once in most case, and at most a handful of times
when not, so validation is essentially linear on the the supergraph edges (which
roughtly correspond to the total number of fields in the supergraph).

The commit also adds a handful of minor of additional performance improvement for
a few things that showed up prominently during profiling. The main one is that
all the assertions in the code where eagerly computing their error message string,
some of which being a tad costly to do and in a hot path. So this commit allows
`assert` to take a closure to generate the error message (which is thus only
generated if the assert fails) and the commit use this in a few sensitive places.