Fix issue where some path were wrongly ignored
The code had an optimization that was trying to avoid using
multiple paths to the same subgraph, but in doing so it
was sometimes excluding more efficient paths (in some cases with
multiple keys in particular).

This commit ensures that those path are now tried and the most
efficient one kept.