Fix TS errors related to FieldSets
We have some overlapping types like FederationType and FederationTypeMetadata
where properties like keys and requires should always be a FieldSet.

This commit simplifies the code a bit and adds assertions+casts where
necessary to ensure these types match as we intend them to.

How these types have been out of sync for so long is surprising. I've
noticed them in the past but it's never been a blocking issue until
I was working on some code with @o0Ignition0o and we ran into these
TS errors.