Finish federation 2 code integration in gateway
This commit ensures the new composition is used in most places,
fixing/updating tests accordingly.

More precisely, after this commit, the fed1 `composeAndValidate`
metohd is is no longer used and removed by this commit. However
the `composeServices` sub-method that the latter method was
calling is still present as a number of validation-specific
tests still rely on it. Updating those tests remains TBD, but
we might want to handle this along the addition of the
legacy mode as at least some of those tests will only apply
to this mode.