Implement condition exclusion when advancing path
The general idea is that if we try to satisfy a condition X for a key
edge, then there is no point in trying to check other key edge with
that same condition. Avoiding this drastically reduce the explored
state in some cases (typically, an entity defined in many subgraph
with the same key).