chore: Split `@apollo/subgraph` package from `@apollo/federation` (#1058)
This commit separates the concerns of running a subgraph from composition,
which the @apollo/federation package is currently responsible for both of. The
new @apollo/subgraph package will be used in the same exact manner for
subgraph developers.

There were some interdependent types, functions, etc. within the federation
package which I had to separate to one or the other. Given that the
@apollo/federation already had an implicit dependency on this new package (for
forward-compat), I chose to migrate the few bits into the new package and deeply
import them from @apollo/federation. I'm not particularly interested in exposing all
of those common bits to the minimal public API of @apollo/subgraph, hence the
nested import.

Other notable updates:
* Docs
* Monorepo infra (TS references, Jest, Lerna)
* Respective READMEs