CHANGELOG update publish prompt
With this commit, Lerna will now pause during the version
step (via the version lifecycle script) in order to allow
the publisher to update CHANGELOGs. This ensures that
CHANGELOG updates are not forgotten, with the added benefit
that they will now also be batched into the publish commit
rather than before or after.