Fix @tag printing in subgraphs
Tag directive usages and definitions weren't being
properly printed by printSubgraphSchema. This was
a copy/paste error which tried to leverage how
printSupergraphSchema printed @tag usages, but it
really just didn't make sense in the context of a
subgraph. We have 2 very different approaches to
reading and printing metadata. The tests were also
incorrect, so fixing the bug triggered a snapshot
update which now reflects the printer is
functioning as expected.