chore: Use router-bridge as a new rust crate and TypeScript project. (#1079)
This is a Rust-JS hybrid package that allows us to invoke some existing JavaScript functions via V8 + deno_core from Rust code.

It's inspired by the existing harmonizer package, but makes this a real TypeScript package so we don't merely need to leverage JSDoc comments for types (which was fraught) and instead uses the TypeScript compiler as typically expected in this monorepo.

The immediate motivation is to expose introspection capabilities via graphql-js to Rust, in order to (as a stop-gap shortcut) as an introspection provider for the Router while we build up the native functionality to do so in Rust via apollo-rs, which will need some time to build that out.

A follow-up motivation – for building this as a general purpose router bridge — will be to land the changes of #812 (JS query planning) on in this router-bridge package, rather than in harmonizer. That's desirable since harmonizer is meant to just be composition for Rover.

A couple notes as to how this differs from harmonizer at the moment — it will probably be worth applying similar changes to the harmonizer package, but those can be follow-ups to this PR:

    The JavaScript invocation can now return anything that implements DeserializeOwned, whereas it previously only returned Strings.
    The runtime.js file remains as just JS as it wasn't happy with assigning things to process and global — which we need to do since we are the runtime environment (and those things are typically provided by Node.js). This work-around definitely suggests there's an inadvertent inclusion of node types, but it didn't seem worth fighting since this (boilerplate) is quite static and doesn't need to change much. In a similar spirit, the tsconfig.json for the router-bridge disables "use strict"; emission. While not normally desirable, this is acceptable because these files still go through rollup (JS) afterward, which adds it anyhow at a higher level.
    Directories have been adjusted to be slightly more intuitive for a Rust/JS hybrid package. In particular:
        TypeScript source files are now in js-src/, rather than src/.
        Rust remains in src/ at the moment, but could theoretically be put in rs-src/ in a later commit.
        dist/ is now js-dist/ and is now emitted by TypeScript compiler, rather than (somewhat confusingly) the rollup bundler.
        The rollup-produced bundles are now in bundled!
    Renamed url_shim to url_polyfill, because that's what it is really!
    Actually has Jest tests wired up to the root TypeScript project that live in router-bridge. This should help ensure compatibility and act as a force-function to keep things aligned and working as the rest of the federation repository evolves.

Lastly, not only is this a language hybrid but this PR is a contribution hybrid — with @o0Ignition0o doing the Rust work and @abernix on the TypeScript. 🎉

Co-authored-by: Jeremy Lempereur jeremy.lempereur@iomentum.com
Co-authored-by: Jesse Rosenberger git@jro.cc