Query planner skips fetches when appropriate
It's reasonable for the query planner to make use of top-level
@skip and @include usages in order to determine whether it can
omit/ignore a FetchNode altogether and prevent an unnecessary call
to a service when possible.

This commit implements this behavior. FetchNodes are appended with
the sum of all "inclusion conditions" of the top level fields.
Each condition has a possible skip and include value, and those
values can be either a literal (true/false) or a variable.

During execution, the executor will reduce all of the conditions
based on literal and variable values into a single "skip or include"
decision, and will either execute or ignore that FetchNode.

A number of edge cases were considered before and during the
implementation of this improvement. The most notable:

  1. Using skip and include on the same selection (always skip
     unless skip: false AND include: true). This adheres to the spec
  2. FragmentSpread and InlineFragment cases. This behavior is
     partially implemented. There are test cases (skipped) to
     demonstrate both ideal behavior and existing limitations of this
     improvement. Resolving these limitations requires some other,
     unrelated changes to the QP impl. details which I consider to be
     out of scope for this change. These cases are also less likely
     to occur and I find the current changeset to be large enough to
     defer solving this edge case for a later time if/when it's
     deemed useful/necessary.