Prevent query planning from running too long
For some queries, it's possible that the possible number of plans is
extremely large. Essentially, every time we have a choice for where we
could resolver a field, the duplicate the number of possible plans,
so this exponentially explode. This commits detect when that's the case
and, if the number of plans is too large, limits the ones we consider
by dropping possible options. Meaning that we give up on producing
the optimal query plan if there is just too many options.

Note that as of this commit, there is very little "intelligence" in
the choice of the options we decide to ignore and we could improve
that. That said, as the query planning algorithm already trim
definitively-ineffecient options when it can, the reducing of
options of this commit doesn't kick it too often and it's probably
good enough for now.