Add schema upgrader
This commit introduces code to automatically migrate (whenever possible)
fed1 schema into fed2 ones.

That code is run automatically during composition on fed1 schema, so
that internally, only fed2 schema are composed. The code however make
sure that composition errors points to the "original" schema, the one
the user inputed, not the automagically migrated one.

The migration code is written is such a way that a separate tool to
convert fed1 schema into fed2 ones could be easily written (and provide
insights into the conversion), but such tool is not part of this commit.

Importantly, as of this commit, we need to be able to identify fed1 vs
fed2 schema, and this commit recognize fed2 schema as those "linking"
to a federation "core" feature with something like:
```
extend schema @link(url: "https://specs.apollo.dev/federation/v2.0/")
```