Update @apollo/core-schema
The latest version of graphql-js made incompatible
changes that broke the @apollo/core-schema errors
which extended GraphQLError.

After updating and releasing a new version of
@apollo/core-schema, we need to do the same here
to get this branch back into a passing state.

Note this change updates the graphql peerDependency
and thus will require a major (0.x) version bump since
it's a breaking change.