Note: this commit is cherry-picked from the future PR #1185. This is a backport fix for @apollo/federation@0.29.0 and @apollo/gateway@0.38.1
fix(composition): Properly merge directive usages during composition (version `0.x`) (#1185)

Fixes a bug in the directive usage merging logic. Unifies duplicated merging logic since it
existed in two very adjacent places.

The bug was in the merging logic itself failing to merge in new keys, only keys which existed
in the map already. For example, in the case that directiveMetadata is populated with an
`external` usage, only `external`s would be merged and an incoming `tag` would be ignored.