feat(gateway): Conditional schema fetch based on ifAfterId (#1152)
The Apollo Uplink endpoint now supports conditional schema fetching via a new
parameter ifAfterId. If this parameter is specified then the endpoint will only
return a schema with an ID that is newer than ifAfterId.

This commit adds the ifAfterId parameter to the Gateway supergraph query and
additionally handles the new case where the schema is unchanged accordingly.