Make the parent getter assert that parent exists
Rather than forcing clients to check or assert that the parent exists, do it in the getter. The rest of the changes are either fallouts of this decision or cleaning up type assertions.

Note that in some cases, we had internal functions start accessing the _parent value directly rather than through the parent getter because they are checking to see if the parent exists while in an intermediate state