Remove undefined and add assert to type getter
All tests pass, but I need to continue to read code to make sure there isn't an existing case that could be a problem.

One definite issue is that if someone were to add a reference to type in the set parent callback, there would definitely be an assertion failure at runtime. Not sure how we feel about that.