Document removed code
The `RESERVED_FIELD_USED` code was previously documented as raise if a
user defined one of the `Query._service` or `Query._entities` field.
But turns out that the code on the 0.x branch can never throw that code
in practice because while there is a pre-composition check throwing
that code, the "normalization" that happens _before_ the
pre-composition checks completely removes the `Query._service` or
`Query._entities` field. Further, the `subgraph` module adds those
fields to user subgraph inconditionally, and those field _are_ displayed
in the schema returned by `_service.sdl`, so it would be incorrect
for composition to reject them (essentially, composition would have no
good way to know if the field have been added by the user or the
`buildSubgraphSchema` method (or equivalent)).

Long story short, current fed2 code essentially preserve the existing
behavior on this, but that does mean we should properly document that
`RESERVED_FIELD_USED` is never thrown anymore.