Implement deno-core required snapshot support for router-bridge (#1292)
Implement deno-core snapshots

deno-core has changed somewhere between 0.86.0 to the current version in use in this crate (0.110.0) and we need to update how we use it to align with those changes. deno-core developers advise using snapshots, which is what the change does (building the v8 snapshot in build.rs and baking the snapshot into the router-bridge binary). It has a beneficial side-effect of speeding up the performance of the v8 instance and opens up further opportunities for improvements in the router down the line.

See #1299 for more details.