Fix: Avoid panic!() s when a deserialization error happens.
So far uncaught runtime errors would trigger a panic on the router-bridge.
Deserialization errors unfortunately would lead to panic's, which we want to avoid.

This commit refactors the way router-bridge errors are raised and propagated.

We now have Two layers of errors (we previously only had one):
- One representing errors raised by the script run (such as GraphQLErrors etc...)
- One representing uncaught errors while interfacing with the JS runtime.

Less panics, and more tests \o/