Prevent field argument collision in diffTypeNodes (#1298)
Previously, arguments were being diff-ed across every field, without tracking
which field the argument belonged to. This made it possible for certain types
to produce false positives, if that type contained 2 fields that both used the
same argument name, but different types on that argument.

Additionally, the `inputValues` entry in the return object would also contain
field diffs for input objects, due to the fact that the InputValueDefinition gets
called on BOTH field arguments, and fields on input objects.

Both these problems have been addressed by diffing the field arguments on
a per-field basis, and by treating fields on input objects the same as fields on
regular types.