Fix merging of fetch groups for @requires
When computing entity fetches for query plans, we keep track for an
entity fetch of the path into it's "parent" group. But when doing so,
we were keeping the "cast" into the entity type if present. Not only
is that unecessary since entity fetches inputs/selections always start
with casting into the proper entity type, it was harmful as it means
that the logic for @require was sometimes not recognizing that it
could merge fetches (due to those casts), sometimes yielding (obviously)
ineffecient plans.