Supports @link to identify fed2 schema
This commit make it so that schema are only considered fed2 if they have
a @link to the federation spec. Otherwise, they are considered fed1
schema and automaticaly upgraded using the `schemaUpgrader` before
composition.

In doing so, this commit allows composition to accept the renaming of
federation directives, which is responsible for a lot of the changes in
this commit, requiring some amount of refactoring.