Fix typings for GraphQLFildResolver to use v16 defaults
The current typings are restrictive for the TArgs field. Users
who wish to use graphql@^15.8.0 will encounter a default of
{ [key: string]: any } type for the args parameter, which will
effectively prevent them from typing their args parameter correctly.

As demonstrated in the thread:
https://community.apollographql.com/t/buildsubgraphschema-typescript-typing-conflict-with-resolver/2716/3

Changing the type of TArgs to any is effectively the same change that
v16 moves to (defaulting to any instead of the above mentioned object-y
type).