Enable @key on list type
Unsure why I added a validation to reject it in the first place, I must
have assume this wasn't supported by federation 1, but it is as far as I
can tell and in any case it works just fine. So this commit removes the
rejection and adds a few tests for the case.