Avoid incomplete subgraph when extracting from supergraph (#1511)
Due to imprecisions of the fed1 join spec, the code handling supergraph
sometimes initially assume a type is in all subgraph to remove that type
later from some subgraphs as it realises it cannot have been in that
subgraph. As we remove types that way, we have to remove references to
it as well, and this was done correctly for object types through
`removeRecursive`, but for union types, `remove` was used instead. This
fix imply ensure `removeRecursive` is used there as well.