Improve handling of invalid enum values
Default values where not properly validated and when an invalid value
was inputed for an enum, we were particularly not handling that
gracefully and were generating invalid graphQL when printing back,
which lead to hard-to-debug errors. This commit fixes both issues
(ensuring we print back syntactically valid graphQL by using a string,
and otherwise throwing a meaningful error).