Fix repetition in messages of composition validation errors
The code that, during validation, avoided considering some
useless/nonsensical paths for root types was hardcoded to `Query` and
was thus not working properly for other root types (mostly `Mutation`).
The result was that a number of trivially-equivalent duplicate paths
were considered, and while this was harmless when validation succeeds,
when there is a legitimate validation error, it made the error message
repeats the same explanation lines once per-subgraph.