Fix issue with assigning @shareable for fed1 migrated schema
The schema upgrader rely on the `sharing` computation on fed1 schema as
a shortcut to decide if a field is part of a key (any key in the
subgraph). The idea is that the only non-external fields are that are shareable by
default are the one that are part of a key, and since fed1 schema don't
have @shareable at all, then the sharing computation does, for fed1
schema, essentially identify fields part of a key.

The only small issue is that the code was assuming that fields that are
part of @provides were always external, which is true for fed2 schema
(which has validation to reject non-external fields in @provides), but
not necessarily for fed1 schema. As a result, the schema upgraded wasn't
adding @shareable for some non-external field even if they needed it
(because they were also provided by another subgraph).