Fix adding descriptions to type extensions
If a schema had both a type definition and a correspdonding type
extension, the printing code was mistakenly printing the description of
the type definition on the type extension as well. But since description
are not syntactically valid on type extension in graphQL, this leads to
syntactic errors.

More precisely, an input schema like:
```graphql
""" I'm a type """
type T {
  a: Int
}

extend type T {
  b: Int
}
```
was (before this patch) mistakenly printed as:
```graphql
""" I'm a type """
type T {
  a: Int
}

""" I'm a type """
extend type T {
  b: Int
}
```
which is syntactically invalid due to the description on the type
extension.

Note: the patch also fix a related additional issue where the presence
of type extension might have the type definition description ignored due
a lack of conditional when setting descriptions in `buildSchema`.